Multi release build
