+++ src/main/java/zmq/io/net/ipc/IpcAddress.java
@@ -1,22 +1,23 @@
 package zmq.io.net.ipc;
 
-import java.io.IOException;
+import java.net.Inet4Address;
+import java.net.Inet6Address;
+import java.net.InetAddress;
+import java.net.InetSocketAddress;
+import java.net.NetworkInterface;
 import java.net.ProtocolFamily;
 import java.net.SocketAddress;
+import java.net.SocketException;
 import java.net.StandardProtocolFamily;
-import java.net.UnixDomainSocketAddress;
-import java.nio.file.Files;
-import java.nio.file.Path;
+import java.util.Enumeration;
 
-import org.zeromq.ZMQException;
-
-import zmq.ZError;
+import zmq.ZMQ;
 import zmq.io.net.Address;
 import zmq.io.net.tcp.TcpAddress;
+import zmq.util.Utils;
 
 public class IpcAddress implements Address.IZAddress
 {
-    // TODO unused?
     public static class IpcAddressMask extends TcpAddress
     {
         public IpcAddressMask(String addr, boolean ipv6)
@@ -30,53 +31,73 @@ public class IpcAddress implements Address.IZAddress
         }
     }
 
-    private String                        name;
-    private final UnixDomainSocketAddress address;
+    public static boolean isIpcAddress(SocketAddress address) {
+        return false;
+    }
+
+    public static String getAddress(SocketAddress address) {
+        throw new IllegalArgumentException("Address is not unix domain socket address.");
+    }
+
+    public static boolean isImplementedViaTcpLoopback()
+    {
+        return true;
+    }
+
+    private String                  name;
+    private final InetSocketAddress address;
 
     public IpcAddress(String addr)
     {
-        // TODO inline?
-        this.address = this.resolve(addr);
+        String[] strings = addr.split(";");
+
+        address = resolve(strings[0]);
+        if (strings.length == 2 && !"".equals(strings[1])) {
+            resolve(strings[1]);
+        }
     }
 
     @Override
     public String toString()
     {
-        // TODO possible?
         if (name == null) {
             return "";
         }
 
-        return "ipc://" + this.address.toString();
+        return "ipc://" + name;
     }
 
     @Override
     public String toString(int port)
     {
-        // TODO why is port in the interface?
+        if ("*".equals(name)) {
+            String suffix = Utils.unhash(port - 10000);
+            return "ipc://" + suffix;
+        }
         return toString();
     }
 
-    private UnixDomainSocketAddress resolve(String name)
+    private InetSocketAddress resolve(String name)
     {
         this.name = name;
 
-        if (!"*".equals(name)) {
-            return UnixDomainSocketAddress.of(name);
+        int hash = name.hashCode();
+        if ("*".equals(name)) {
+            hash = 0;
+        }
+        else {
+            if (hash < 0) {
+                hash = -hash;
+            }
+            hash = hash % 55536;
+            hash += 10000;
         }
 
-        try {
-            Path temp = Files.createTempFile("zmq-", ".sock");
-            Files.delete(temp);
-            return UnixDomainSocketAddress.of(temp);
-        }
-        catch (IOException e) {
-            throw new ZMQException(e.getMessage(), ZError.EADDRNOTAVAIL, e);
-        }
+        return new InetSocketAddress(findAddress(), hash);
     }
 
     @Override
-    public UnixDomainSocketAddress address()
+    public SocketAddress address()
     {
         return address;
     }
@@ -84,6 +105,27 @@ public class IpcAddress implements Address.IZAddress
     @Override
     public ProtocolFamily family()
     {
-        return StandardProtocolFamily.UNIX;
+        return StandardProtocolFamily.INET;
+    }
+
+    private InetAddress findAddress()
+    {
+        Class<? extends InetAddress> addressClass = ZMQ.PREFER_IPV6 ? Inet6Address.class : Inet4Address.class;
+        try {
+            for (Enumeration<NetworkInterface> interfaces = NetworkInterface.getNetworkInterfaces(); interfaces
+                    .hasMoreElements();) {
+                NetworkInterface net = interfaces.nextElement();
+                for (Enumeration<InetAddress> addresses = net.getInetAddresses(); addresses.hasMoreElements();) {
+                    InetAddress inetAddress = addresses.nextElement();
+                    if (inetAddress.isLoopbackAddress() && addressClass.isInstance(inetAddress)) {
+                        return inetAddress;
+                    }
+                }
+            }
+        }
+        catch (SocketException e) {
+            throw new IllegalArgumentException(e);
+        }
+        throw new IllegalArgumentException("no address found " + (ZMQ.PREFER_IPV6 ? "IPV6" : "IPV4") + "local");
     }
 }
